requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
gunicorn==21.2.0
packaging==23.1
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.7.1
